holy... how did this end up being so big...???
Well, this makes user_api pretty much complete except for security stuff

I still need to check through all the page files because I'm sure there are still pages that aren't calling user_api functions.  In fact there are user_api functions that haven't been written yet that I'm pretty sure will need to be...

But there ya go... it's better than it was before :)


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1384 f5dc347c-c33d-0410-90a0-b07cc1902cb9

+++ core/database_api.php
@@ -6,7 +6,7 @@
 	# See the files README and LICENSE for details
 
 	# --------------------------------------------------------
-	# $Id: database_api.php,v 1.5 2002-08-27 04:26:42 jfitzell Exp $
+	# $Id: database_api.php,v 1.6 2002-08-29 02:56:23 jfitzell Exp $
 	# --------------------------------------------------------
 
 	###########################################################################
@@ -161,7 +161,7 @@
 	# --------------------
 	# prepare a boolean before DB insertion
 	function db_prepare_bool( $p_bool ) {
-		return (bool)$p_bool;
+		return (int)(bool)$p_bool;
 	}
 
 	# --------------------
